requirements-orchestration.txt
pandas
numpy
psycopg[binary]
great-expectations>=0.18,<1
apache-airflow>=2.10,<3
